refactor(kor): move kor.h into adatok.h
