--- KOR.H
@@ -1,10 +1,0 @@
-
-struct rigidbody {
-    double rotation;
-    double angular_velocity;
-    double radius;
-    double mass;
-    double inertia; // Moment of inertia
-    vekt2 r;
-    vekt2 v;
-};
